+++ R/format_parameters.R
@@ -68,7 +68,7 @@ format_parameters.parameters_model <- function(model, ...) {
 
 
 .format_parameter_default <- function(model, effects = "fixed", brackets = c("[", "]"), ...) {
-  original_names <- names <- insight::find_parameters(model, effects = effects, flatten = TRUE)
+  original_names <- parameter_names <- insight::find_parameters(model, effects = effects, flatten = TRUE)
 
   # save some time, if model info is passed as argument
   dot_args <- list(...)
@@ -85,7 +85,7 @@ format_parameters.parameters_model <- function(model, ...) {
 
   # quick fix, for multivariate response models, we use
   # info from first model only
-  if (insight::is_multivariate(model) && !"is_zero_inflated" %in% names(info)) {
+  if (insight::is_multivariate(model) && !"is_zero_inflated" %in% names(info) && !inherits(model, c("vgam", "vglm"))) {
     info <- info[[1]]
   }
 
@@ -100,19 +100,19 @@ format_parameters.parameters_model <- function(model, ...) {
 
   # special handling hurdle- and zeroinfl-models ---------------------
   if (isTRUE(info$is_zero_inflated) || isTRUE(info$is_hurdle)) {
-    names <- gsub("^(count_|zero_)", "", names)
+    parameter_names <- gsub("^(count_|zero_)", "", parameter_names)
     types$Parameter <- gsub("^(count_|zero_)", "", types$Parameter)
   }
 
   # special handling polr ---------------------
   if (inherits(model, "polr")) {
     original_names <- gsub("Intercept: ", "", original_names, fixed = TRUE)
-    names <- gsub("Intercept: ", "", names, fixed = TRUE)
+    parameter_names <- gsub("Intercept: ", "", parameter_names, fixed = TRUE)
   }
 
   # special handling bracl ---------------------
   if (inherits(model, "bracl")) {
-    names <- gsub("(.*):(.*)", "\\2", names)
+    parameter_names <- gsub("(.*):(.*)", "\\2", parameter_names)
   }
 
   # special handling DirichletRegModel ---------------------
@@ -121,11 +121,11 @@ format_parameters.parameters_model <- function(model, ...) {
     cf <- stats::coef(model)
     if (model$parametrization == "common") {
       pattern <- paste0("(", paste(model$varnames, collapse = "|"), ")\\.(.*)")
-      dirich_names <- names <- gsub(pattern, "\\2", names(unlist(cf)))
+      dirich_names <- parameter_names <- gsub(pattern, "\\2", names(unlist(cf)))
     } else {
-      dirich_names <- names <- gsub("(.*)\\.(.*)\\.(.*)", "\\3", names(unlist(cf)))
+      dirich_names <- parameter_names <- gsub("(.*)\\.(.*)\\.(.*)", "\\3", names(unlist(cf)))
     }
-    original_names <- names
+    original_names <- parameter_names
     if (!is.null(dirich_names)) {
       types$Parameter <- dirich_names
     }
@@ -133,25 +133,14 @@ format_parameters.parameters_model <- function(model, ...) {
 
 
   # remove "as.factor()", "log()" etc. from parameter names
-  names <- .clean_parameter_names(names)
+  parameter_names <- .clean_parameter_names(parameter_names)
 
 
   for (i in seq_len(nrow(types))) {
     name <- types$Parameter[i]
 
-    # No interaction
-    if (!types$Type[i] %in% c("interaction", "nested", "simple")) {
-      type <- types[i, ]
-      names[i] <- .format_parameter(
-        name,
-        variable = type$Variable,
-        type = type$Type,
-        level = type$Level,
-        brackets = brackets
-      )
-
+    if (types$Type[i] %in% c("interaction", "nested", "simple")) {
       # Interaction or nesting
-    } else {
       components <- unlist(strsplit(name, ":", fixed = TRUE), use.names = FALSE)
       is_nested <- types$Type[i] == "nested"
       is_simple <- types$Type[i] == "simple"
@@ -186,20 +175,30 @@ format_parameters.parameters_model <- function(model, ...) {
           )
         }
       }
-      names[i] <- .format_interaction(
+      parameter_names[i] <- .format_interaction(
         components = components,
         type = types[i, "Type"],
         is_nested = is_nested,
         is_simple = is_simple,
         ...
       )
+    } else {
+      # No interaction
+      type <- types[i, ]
+      parameter_names[i] <- .format_parameter(
+        name,
+        variable = type$Variable,
+        type = type$Type,
+        level = type$Level,
+        brackets = brackets
+      )
     }
   }
 
   # do some final formatting, like replacing underscores or dots with whitespace.
-  names <- gsub("(\\.|_)(?![^\\[]*\\])", " ", names, perl = TRUE)
+  parameter_names <- gsub("(\\.|_)(?![^\\[]*\\])", " ", parameter_names, perl = TRUE)
   # remove double spaces
-  names <- gsub("  ", " ", names, fixed = TRUE)
+  parameter_names <- gsub("  ", " ", parameter_names, fixed = TRUE)
 
   # "types$Parameter" here is cleaned, i.e. patterns like "log()", "as.factor()"
   # etc. are removed. However, these patterns are needed in "format_table()",
@@ -208,8 +207,8 @@ format_parameters.parameters_model <- function(model, ...) {
   # so output will be NA resp. blank fields... Thus, I think we should use
   # the original parameter-names here.
 
-  names(names) <- original_names # types$Parameter
-  names
+  names(parameter_names) <- original_names # types$Parameter
+  parameter_names
 }
 
 
@@ -345,9 +344,9 @@ format_parameters.parameters_model <- function(model, ...) {
 #' @keywords internal
 .format_ordered <- function(degree, brackets = c("[", "]")) {
   switch(degree,
-    `.L` = paste0(brackets[1], "linear", brackets[2]),
-    `.Q` = paste0(brackets[1], "quadratic", brackets[2]),
-    `.C` = paste0(brackets[1], "cubic", brackets[2]),
+    .L = paste0(brackets[1], "linear", brackets[2]),
+    .Q = paste0(brackets[1], "quadratic", brackets[2]),
+    .C = paste0(brackets[1], "cubic", brackets[2]),
     paste0(
       brackets[1],
       parameters::format_order(as.numeric(gsub("^", "", degree, fixed = TRUE)), textual = FALSE),
@@ -362,7 +361,7 @@ format_parameters.parameters_model <- function(model, ...) {
 # replace pretty names with value labels, when present ---------------
 
 .format_value_labels <- function(params, model = NULL) {
-  labels <- NULL
+  pretty_labels <- NULL
   if (is.null(model)) {
     model <- .get_object(params)
   }
@@ -391,9 +390,9 @@ format_parameters.parameters_model <- function(model, ...) {
         out <- attr(vec, "label", exact = TRUE)
       }
       if (is.null(out)) {
-        return(i)
+        i
       } else {
-        return(out)
+        out
       }
     })
 
@@ -407,7 +406,7 @@ format_parameters.parameters_model <- function(model, ...) {
 
     # name elements
     names(lbs) <- names(preds) <- colnames(mf)
-    labels <- .safe(stats::setNames(
+    pretty_labels <- .safe(stats::setNames(
       unlist(lbs, use.names = FALSE),
       unlist(preds, use.names = FALSE)
     ))
@@ -416,7 +415,7 @@ format_parameters.parameters_model <- function(model, ...) {
     pn <- attributes(params)$pretty_names
     # replace former pretty names with labels, if we have any labels
     # (else, default pretty names are returned)
-    if (!is.null(labels)) {
+    if (!is.null(pretty_labels)) {
       # check if we have any interactions, and if so, create combined labels
       interactions <- pn[grepl(":", names(pn), fixed = TRUE)]
       if (length(interactions)) {
@@ -425,23 +424,23 @@ format_parameters.parameters_model <- function(model, ...) {
           # extract single coefficient names from interaction term
           out <- unlist(strsplit(i, ":", fixed = TRUE))
           # combine labels
-          labs <- c(labs, paste0(sapply(out, function(l) labels[l]), collapse = " * "))
+          labs <- c(labs, paste0(sapply(out, function(l) pretty_labels[l]), collapse = " * "))
         }
         # add interaction terms to labels string
         names(labs) <- names(interactions)
-        labels <- c(labels, labs)
+        pretty_labels <- c(pretty_labels, labs)
       }
       # make sure "invalid" labels are ignored
-      common_labels <- intersect(names(labels), names(pn))
-      pn[common_labels] <- labels[common_labels]
+      common_labels <- intersect(names(pretty_labels), names(pn))
+      pn[common_labels] <- pretty_labels[common_labels]
     }
-    labels <- pn
+    pretty_labels <- pn
   }
 
   # missing labels return original parameter name (e.g., variance components in mixed models)
   out <- stats::setNames(params$Parameter, params$Parameter)
-  labels <- labels[names(labels) %in% params$Parameter]
-  out[match(names(labels), params$Parameter)] <- labels
+  pretty_labels <- pretty_labels[names(pretty_labels) %in% params$Parameter]
+  out[match(names(pretty_labels), params$Parameter)] <- pretty_labels
 
   out
 }