+++ R/utils_pca_efa.R
@@ -107,7 +107,7 @@ summary.parameters_efa <- function(object, ...) {
 
 
   x <- as.data.frame(t(x[, cols]))
-  x <- cbind(data.frame("Parameter" = row.names(x), stringsAsFactors = FALSE), x)
+  x <- cbind(data.frame(Parameter = row.names(x), stringsAsFactors = FALSE), x)
   names(x) <- c("Parameter", attributes(object)$summary$Component)
   row.names(x) <- NULL
 
@@ -149,40 +149,40 @@ predict.parameters_efa <- function(object,
                                    verbose = TRUE,
                                    ...) {
   attri <- attributes(object)
+
+  # handle if no data is provided
   if (is.null(newdata)) {
+    # check if we have scores attribute - these will be returned directly
     if ("scores" %in% names(attri)) {
       out <- as.data.frame(attri$scores)
       if (isTRUE(keep_na)) {
         out <- .merge_na(object, out, verbose)
       }
+    } else if ("dataset" %in% names(attri)) {
+      # if we have data, use that for prediction
+      d <- attri$data_set
+      d <- d[vapply(d, is.numeric, logical(1))]
+      out <- as.data.frame(stats::predict(attri$model, newdata = d))
     } else {
-      if ("dataset" %in% names(attri)) {
-        out <- as.data.frame(stats::predict(attri$model, data = attri$dataset))
-      } else {
-        insight::format_error(
-          "Could not retrieve data nor model. Please report an issue on {.url https://github.com/easystats/parameters/issues}."
-        )
-      }
+      insight::format_error(
+        "Could not retrieve data nor model. Please report an issue on {.url https://github.com/easystats/parameters/issues}." # nolint
+      )
     }
+  } else if (inherits(attri$model, "spca")) {
+    # https://github.com/erichson/spca/issues/7
+    newdata <- newdata[names(attri$model$center)]
+    if (attri$standardize) {
+      newdata <- sweep(newdata, MARGIN = 2, STATS = attri$model$center, FUN = "-", check.margin = TRUE)
+      newdata <- sweep(newdata, MARGIN = 2, STATS = attri$model$scale, FUN = "/", check.margin = TRUE)
+    }
+    out <- as.matrix(newdata) %*% as.matrix(attri$model$loadings)
+    out <- stats::setNames(as.data.frame(out), paste0("Component", seq_len(ncol(out))))
+  } else if (inherits(attri$model, c("psych", "fa", "principal"))) {
+    out <- as.data.frame(stats::predict(attri$model, data = newdata, ...))
   } else {
-    if (inherits(attri$model, c("psych", "fa"))) {
-      # Clean-up newdata (keep only the variables used in the model)
-      newdata <- newdata[names(attri$model$complexity)] # assuming "complexity" info is there
-      # psych:::predict.fa(object, data)
-      out <- as.data.frame(stats::predict(attri$model, data = newdata))
-    } else if (inherits(attri$model, "spca")) {
-      # https://github.com/erichson/spca/issues/7
-      newdata <- newdata[names(attri$model$center)]
-      if (attri$standardize) {
-        newdata <- sweep(newdata, MARGIN = 2, STATS = attri$model$center, FUN = "-", check.margin = TRUE)
-        newdata <- sweep(newdata, MARGIN = 2, STATS = attri$model$scale, FUN = "/", check.margin = TRUE)
-      }
-      out <- as.matrix(newdata) %*% as.matrix(attri$model$loadings)
-      out <- stats::setNames(as.data.frame(out), paste0("Component", seq_len(ncol(out))))
-    } else {
-      out <- as.data.frame(stats::predict(attri$model, newdata = newdata, ...))
-    }
+    out <- as.data.frame(stats::predict(attri$model, newdata = newdata, ...))
   }
+
   if (!is.null(names)) {
     names(out)[seq_along(names)] <- names
   }
@@ -339,19 +339,17 @@ print.parameters_omega_summary <- function(x, ...) {
     } else {
       table_caption <- c(sprintf("# Loadings from %s (no rotation)", method), "blue")
     }
+  } else if (format == "markdown") {
+    table_caption <- sprintf("Rotated loadings from %s (%s-rotation)", method, rotation_name)
   } else {
-    if (format == "markdown") {
-      table_caption <- sprintf("Rotated loadings from %s (%s-rotation)", method, rotation_name)
-    } else {
-      table_caption <- c(sprintf("# Rotated loadings from %s (%s-rotation)", method, rotation_name), "blue")
-    }
+    table_caption <- c(sprintf("# Rotated loadings from %s (%s-rotation)", method, rotation_name), "blue")
   }
 
   # footer
-  if (!is.null(attributes(x)$type)) {
-    footer <- c(.text_components_variance(x, sep = ifelse(format == "markdown", "", "\n")), "yellow")
-  } else {
+  if (is.null(attributes(x)$type)) {
     footer <- NULL
+  } else {
+    footer <- c(.text_components_variance(x, sep = ifelse(format == "markdown", "", "\n")), "yellow")
   }
 
   insight::export_table(
@@ -381,48 +379,48 @@ print.parameters_omega_summary <- function(x, ...) {
   }
 
   if (type == "cluster") {
-    summary <- as.data.frame(x)
+    cluster_summary <- as.data.frame(x)
     variance <- attributes(x)$variance * 100
   } else {
-    summary <- attributes(x)$summary
-    variance <- max(summary$Variance_Cumulative) * 100
+    cluster_summary <- attributes(x)$summary
+    variance <- max(cluster_summary$Variance_Cumulative) * 100
   }
 
-  if (nrow(summary) == 1) {
-    text <- paste0("The unique ", type)
+  if (nrow(cluster_summary) == 1) {
+    text_variance <- paste0("The unique ", type)
   } else {
-    text <- paste0("The ", nrow(summary), " ", type, "s")
+    text_variance <- paste0("The ", nrow(cluster_summary), " ", type, "s")
   }
 
   # rotation
   if (!is.null(attributes(x)$rotation) && attributes(x)$rotation != "none") {
-    text <- paste0(text, " (", attributes(x)$rotation, " rotation)")
+    text_variance <- paste0(text_variance, " (", attributes(x)$rotation, " rotation)")
   }
 
 
-  text <- paste0(
-    text,
+  text_variance <- paste0(
+    text_variance,
     " accounted for ",
     sprintf("%.2f", variance),
     "% of the total variance of the original data"
   )
 
-  if (type == "cluster" || nrow(summary) == 1) {
-    text <- paste0(text, ".")
+  if (type == "cluster" || nrow(cluster_summary) == 1) {
+    text_variance <- paste0(text_variance, ".")
   } else {
-    text <- paste0(
-      text,
+    text_variance <- paste0(
+      text_variance,
       " (",
-      paste0(summary$Component,
+      paste0(cluster_summary$Component,
         " = ",
-        sprintf("%.2f", summary$Variance * 100),
+        sprintf("%.2f", cluster_summary$Variance * 100),
         "%",
         collapse = ", "
       ),
       ")."
     )
   }
-  paste0(sep, text, sep)
+  paste0(sep, text_variance, sep)
 }
 
 
@@ -482,8 +480,8 @@ sort.parameters_pca <- sort.parameters_efa
     }
   }
 
-  order <- row.names(x)
-  loadings <- loadings[as.numeric(as.character(order)), ] # Arrange by max
+  row_order <- row.names(x)
+  loadings <- loadings[as.numeric(as.character(row_order)), ] # Arrange by max
   row.names(loadings) <- NULL
 
   loadings